Added a service for each availability search step
