+++ Api/HappyTravel.Edo.Api.xml
@@ -1819,22 +1819,47 @@
                 Payfort tokenization url
             </summary>
         </member>
-        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.AvailabilitySearchState.Id">
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.ProviderAvailabilitySearchState.Id">
             <summary>
             Search id
             </summary>
         </member>
-        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.AvailabilitySearchState.TaskState">
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.ProviderAvailabilitySearchState.TaskState">
             <summary>
             Search task state
             </summary>
         </member>
-        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.AvailabilitySearchState.ResultCount">
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.ProviderAvailabilitySearchState.ResultCount">
             <summary>
             Result count
             </summary>
         </member>
-        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.AvailabilitySearchState.Error">
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.ProviderAvailabilitySearchState.Error">
+            <summary>
+            Error message. Filled only for failed tasks
+            </summary>
+        </member>
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1.AvailabilitySearchState.Id">
+            <summary>
+            Search id
+            </summary>
+        </member>
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1.AvailabilitySearchState.TaskState">
+            <summary>
+            Search task state
+            </summary>
+        </member>
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1.AvailabilitySearchState.ProviderStates">
+            <summary>
+            Provider search states
+            </summary>
+        </member>
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1.AvailabilitySearchState.ResultCount">
+            <summary>
+            Result count
+            </summary>
+        </member>
+        <member name="P:HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1.AvailabilitySearchState.Error">
             <summary>
             Error message. Filled only for failed tasks
             </summary>
