+++ Api/Infrastructure/ServiceCollectionExtensions.cs
@@ -20,6 +20,9 @@ using HappyTravel.Edo.Api.Services.Accommodations;
 using HappyTravel.Edo.Api.Services.Accommodations.Availability;
 using HappyTravel.Edo.Api.Services.Accommodations.Bookings;
 using HappyTravel.Edo.Api.AdministratorServices;
+using HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1;
+using HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step2;
+using HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step3;
 using HappyTravel.Edo.Api.Services.Accommodations.Mappings;
 using HappyTravel.Edo.Api.Services.Agents;
 using HappyTravel.Edo.Api.Services.CodeProcessors;
@@ -447,7 +450,9 @@ namespace HappyTravel.Edo.Api.Infrastructure
             services.AddTransient<IBookingDocumentsService, BookingDocumentsService>();
             services.AddTransient<IBookingAuditLogService, BookingAuditLogService>();
             services.AddTransient<IDataProviderFactory, DataProviderFactory>();
-            services.AddTransient<IAvailabilityService, AvailabilityService>();
+            services.AddTransient<IFirstStepAvailabilitySearchService, FirstStepAvailabilitySearchService>();
+            services.AddTransient<ISecondStepAvailabilitySearchService, SecondStepAvailabilitySearchService>();
+            services.AddTransient<IThirdStepAvailabilitySearchService, ThirdStepAvailabilitySearchService>();
             services.AddTransient<IBookingService, BookingService>();
             services.AddTransient<IBookingsProcessingService, BookingsProcessingService>();
             services.AddTransient<IProviderRouter, ProviderRouter>();
