--- Api/Services/Accommodations/Availability/AvailabilityService.cs
@@ -1,120 +1,0 @@
-using System;
-using System.Threading.Tasks;
-using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Models.Accommodations;
-using HappyTravel.Edo.Api.Models.Markups;
-using HappyTravel.Edo.Api.Services.Agents;
-using HappyTravel.Edo.Api.Services.Connectors;
-using HappyTravel.Edo.Common.Enums;
-using HappyTravel.EdoContracts.Accommodations;
-using Microsoft.AspNetCore.Mvc;
-
-namespace HappyTravel.Edo.Api.Services.Accommodations.Availability
-{
-    public class AvailabilityService : IAvailabilityService
-    {
-        public AvailabilityService(IAgentContextService agentContextService,
-            IPriceProcessor priceProcessor,
-            IAvailabilityResultsCache availabilityResultsCache,
-            IProviderRouter providerRouter)
-        {
-            _agentContextService = agentContextService;
-            _priceProcessor = priceProcessor;
-            _availabilityResultsCache = availabilityResultsCache;
-            _providerRouter = providerRouter;
-        }
-
-
-        public async Task<Result<ProviderData<SingleAccommodationAvailabilityDetails>, ProblemDetails>> GetAvailable(DataProviders dataProvider,
-            string accommodationId, string availabilityId,
-            string languageCode)
-        {
-            var agent = await _agentContextService.GetAgent();
-
-            return await ExecuteRequest()
-                .Bind(ConvertCurrencies)
-                .Map(ApplyMarkups)
-                .Map(AddProviderData);
-
-
-            Task<Result<SingleAccommodationAvailabilityDetails, ProblemDetails>> ExecuteRequest()
-                => _providerRouter.GetAvailable(dataProvider, accommodationId, availabilityId, languageCode);
-
-
-            Task<Result<SingleAccommodationAvailabilityDetails, ProblemDetails>> ConvertCurrencies(SingleAccommodationAvailabilityDetails availabilityDetails)
-                => _priceProcessor.ConvertCurrencies(agent, availabilityDetails, AvailabilityResultsExtensions.ProcessPrices, AvailabilityResultsExtensions.GetCurrency);
-
-
-            Task<DataWithMarkup<SingleAccommodationAvailabilityDetails>> ApplyMarkups(SingleAccommodationAvailabilityDetails response) 
-                => _priceProcessor.ApplyMarkups(agent, response, AvailabilityResultsExtensions.ProcessPrices);
-
-
-            ProviderData<SingleAccommodationAvailabilityDetails> AddProviderData(DataWithMarkup<SingleAccommodationAvailabilityDetails> availabilityDetails)
-                => ProviderData.Create(dataProvider, availabilityDetails.Data);
-        }
-
-
-        public async Task<Result<ProviderData<SingleAccommodationAvailabilityDetailsWithDeadline?>, ProblemDetails>> GetExactAvailability(
-            DataProviders dataProvider, string availabilityId, Guid roomContractSetId, string languageCode)
-        {
-            var agent = await _agentContextService.GetAgent();
-
-            return await ExecuteRequest()
-                .Bind(ConvertCurrencies)
-                .Map(ApplyMarkups)
-                .Tap(SaveToCache)
-                .Map(AddProviderData);
-
-
-            Task<Result<SingleAccommodationAvailabilityDetailsWithDeadline?, ProblemDetails>> ExecuteRequest()
-                => _providerRouter.GetExactAvailability(dataProvider, availabilityId, roomContractSetId, languageCode);
-
-
-            Task<Result<SingleAccommodationAvailabilityDetailsWithDeadline?, ProblemDetails>> ConvertCurrencies(SingleAccommodationAvailabilityDetailsWithDeadline? availabilityDetails) => _priceProcessor.ConvertCurrencies(agent,
-                availabilityDetails,
-                AvailabilityResultsExtensions.ProcessPrices,
-                AvailabilityResultsExtensions.GetCurrency);
-
-
-            Task<DataWithMarkup<SingleAccommodationAvailabilityDetailsWithDeadline?>>
-                ApplyMarkups(SingleAccommodationAvailabilityDetailsWithDeadline? response)
-                => _priceProcessor.ApplyMarkups(agent, response, AvailabilityResultsExtensions.ProcessPrices);
-
-
-            Task SaveToCache(DataWithMarkup<SingleAccommodationAvailabilityDetailsWithDeadline?> responseWithDeadline)
-            {
-                if(!responseWithDeadline.Data.HasValue)
-                    return Task.CompletedTask;
-                
-                return _availabilityResultsCache.Set(dataProvider, DataWithMarkup.Create(responseWithDeadline.Data.Value, 
-                    responseWithDeadline.Policies));
-            }
-
-
-            ProviderData<SingleAccommodationAvailabilityDetailsWithDeadline?> AddProviderData(
-                DataWithMarkup<SingleAccommodationAvailabilityDetailsWithDeadline?> availabilityDetails)
-                => ProviderData.Create(dataProvider, availabilityDetails.Data);
-        }
-
-
-        public Task<Result<ProviderData<DeadlineDetails>, ProblemDetails>> GetDeadlineDetails(
-            DataProviders dataProvider, string availabilityId, Guid roomContractSetId, string languageCode)
-        {
-            return GetDeadline()
-                .Map(AddProviderData);
-
-            Task<Result<DeadlineDetails, ProblemDetails>> GetDeadline() => _providerRouter.GetDeadline(dataProvider,
-                availabilityId,
-                roomContractSetId, languageCode);
-
-            ProviderData<DeadlineDetails> AddProviderData(DeadlineDetails deadlineDetails)
-                => ProviderData.Create(dataProvider, deadlineDetails);
-        }
-
-
-        private readonly IAvailabilityResultsCache _availabilityResultsCache;
-        private readonly IAgentContextService _agentContextService;
-        private readonly IProviderRouter _providerRouter;
-        private readonly IPriceProcessor _priceProcessor;
-    }
-}
