+++ Api/Services/Accommodations/Availability/Steps/Step1/IAvailabilitySearchScheduler.cs
@@ -4,7 +4,7 @@ using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Models.Availabilities;
 
-namespace HappyTravel.Edo.Api.Services.Accommodations.Availability
+namespace HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step1
 {
     public interface IAvailabilitySearchScheduler
     {
