+++ Api/Services/Accommodations/Availability/Steps/Step3/IThirdStepAvailabilitySearchService.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Threading.Tasks;
+using CSharpFunctionalExtensions;
+using HappyTravel.Edo.Api.Models.Accommodations;
+using HappyTravel.Edo.Api.Models.Agents;
+using HappyTravel.Edo.Common.Enums;
+using HappyTravel.EdoContracts.Accommodations;
+using Microsoft.AspNetCore.Mvc;
+
+namespace HappyTravel.Edo.Api.Services.Accommodations.Availability.Steps.Step3
+{
+    public interface IThirdStepAvailabilitySearchService
+    {
+        Task<Result<ProviderData<SingleAccommodationAvailabilityDetailsWithDeadline?>, ProblemDetails>> GetExactAvailability(
+            DataProviders dataProvider, string availabilityId, Guid roomContractSetId, AgentContext agent, string languageCode);
+    }
+}
