+++ HappyTravel.Edo.UnitTests/Utility/AvailabilityStorageUtils.cs
@@ -41,9 +41,6 @@ namespace HappyTravel.Edo.UnitTests.Utility
                     
             
             return new AvailabilityStorage(distributedFlowMock.Object,
-                Mock.Of<IMemoryFlow>(), 
-                new DefaultDateTimeProvider(), 
-                Mock.Of<IAccommodationDuplicatesService>(), 
                 providerOptions);
         }
     }
